pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "httpdbg"
authors = [
    {name = "cle-b", email = "cle@tictac.pm"},
]
description="A very simple tool to debug HTTP(S) client requests"
readme="README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Testing",
    "Topic :: Utilities",
]
dynamic = ["version"]

[project.scripts]
pyhttpdbg = "httpdbg.__main__:pyhttpdbg_entry_point"

[tool.setuptools]
packages = ["httpdbg", "httpdbg.hooks", "httpdbg.webapp"]

[tool.setuptools.package-data]
httpdbg = ["webapp/static/*", "webapp/static/icons/*"]

[tool.setuptools.dynamic]
version = {attr = "httpdbg.__version__"}

[tool.mypy]
no_implicit_optional = false
disable_error_code= ["method-assign",]
files = ["httpdbg"]

[project.optional-dependencies]
pytest = ["pytest-httpdbg>=0.6.0"]

[project.urls]
Source = "https://github.com/cle-b/httpdbg/"
Documentation = "https://httpdbg.readthedocs.io/"
